+++ src/main/java/nextstep/global/exception/AuthenticationException.java
@@ -1,4 +1,4 @@
-package nextstep.member;
+package nextstep.global.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
