+++ src/main/java/nextstep/member/application/TokenService.java
@@ -1,6 +1,6 @@
 package nextstep.member.application;
 
-import nextstep.member.AuthenticationException;
+import nextstep.global.exception.AuthenticationException;
 import nextstep.member.application.dto.TokenResponse;
 import nextstep.member.domain.Member;
 import org.springframework.stereotype.Service;
